chateuler/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
jiter==0.5.0
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
jwt==1.3.1
langchain==0.2.15
langchain-core==0.2.37
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.108
multidict==6.0.5
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.21
sqlmodel==0.0.22
starlette==0.37.2
strawberry-graphql==0.237.3
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
yarl==1.9.7
